+++ Aulas-Angular/app-modules/.angular/cache/17.3.8/vite/deps/@angular_common_http.js
@@ -1,0 +1,83 @@
+import {
+  FetchBackend,
+  HTTP_INTERCEPTORS,
+  HTTP_ROOT_INTERCEPTOR_FNS,
+  HttpBackend,
+  HttpClient,
+  HttpClientJsonpModule,
+  HttpClientModule,
+  HttpClientXsrfModule,
+  HttpContext,
+  HttpContextToken,
+  HttpErrorResponse,
+  HttpEventType,
+  HttpFeatureKind,
+  HttpHandler,
+  HttpHeaderResponse,
+  HttpHeaders,
+  HttpInterceptorHandler,
+  HttpParams,
+  HttpRequest,
+  HttpResponse,
+  HttpResponseBase,
+  HttpStatusCode,
+  HttpUrlEncodingCodec,
+  HttpXhrBackend,
+  HttpXsrfTokenExtractor,
+  JsonpClientBackend,
+  JsonpInterceptor,
+  PRIMARY_HTTP_BACKEND,
+  provideHttpClient,
+  withFetch,
+  withHttpTransferCache,
+  withInterceptors,
+  withInterceptorsFromDi,
+  withJsonpSupport,
+  withNoXsrfProtection,
+  withRequestsMadeViaParent,
+  withXsrfConfiguration
+} from "./chunk-BBUMVJIZ.js";
+import "./chunk-V4I6KYGB.js";
+import "./chunk-CHRFX2HE.js";
+import "./chunk-R7GQRDZ6.js";
+export {
+  FetchBackend,
+  HTTP_INTERCEPTORS,
+  HttpBackend,
+  HttpClient,
+  HttpClientJsonpModule,
+  HttpClientModule,
+  HttpClientXsrfModule,
+  HttpContext,
+  HttpContextToken,
+  HttpErrorResponse,
+  HttpEventType,
+  HttpFeatureKind,
+  HttpHandler,
+  HttpHeaderResponse,
+  HttpHeaders,
+  HttpParams,
+  HttpRequest,
+  HttpResponse,
+  HttpResponseBase,
+  HttpStatusCode,
+  HttpUrlEncodingCodec,
+  HttpXhrBackend,
+  HttpXsrfTokenExtractor,
+  JsonpClientBackend,
+  JsonpInterceptor,
+  provideHttpClient,
+  withFetch,
+  withInterceptors,
+  withInterceptorsFromDi,
+  withJsonpSupport,
+  withNoXsrfProtection,
+  withRequestsMadeViaParent,
+  withXsrfConfiguration,
+  HTTP_ROOT_INTERCEPTOR_FNS as ɵHTTP_ROOT_INTERCEPTOR_FNS,
+  HttpInterceptorHandler as ɵHttpInterceptingHandler,
+  HttpInterceptorHandler as ɵHttpInterceptorHandler,
+  PRIMARY_HTTP_BACKEND as ɵPRIMARY_HTTP_BACKEND,
+  withHttpTransferCache as ɵwithHttpTransferCache
+};
+//# sourceMappingURL=@angular_common_http.js.map
