+++ Aulas-Angular/app-modules/.angular/cache/17.3.8/vite/deps/@angular_platform-browser.js
@@ -35,7 +35,8 @@ import {
   provideProtractorTestingSupport,
   withHttpTransferCacheOptions,
   withNoHttpTransferCache
-} from "./chunk-4NJKIFMT.js";
+} from "./chunk-ZN4ZQIKF.js";
+import "./chunk-BBUMVJIZ.js";
 import {
   getDOM
 } from "./chunk-V4I6KYGB.js";
