+++ Aulas-Angular/app-modules/.angular/cache/17.3.8/vite/deps/@angular_router.js
@@ -1,6 +1,7 @@
 import {
   Title
-} from "./chunk-4NJKIFMT.js";
+} from "./chunk-ZN4ZQIKF.js";
+import "./chunk-BBUMVJIZ.js";
 import {
   DOCUMENT,
   HashLocationStrategy,
