+++ Aulas-Angular/app-modules/.angular/cache/17.3.8/vite/deps/_metadata.json
@@ -1,73 +1,79 @@
 {
-  "hash": "77171b8d",
-  "configHash": "bc5ce39b",
+  "hash": "25d36c58",
+  "configHash": "2499f8b4",
   "lockfileHash": "63204a38",
-  "browserHash": "b3bc6563",
+  "browserHash": "266173bb",
   "optimized": {
     "@angular/common": {
       "src": "../../../../../node_modules/@angular/common/fesm2022/common.mjs",
       "file": "@angular_common.js",
-      "fileHash": "3cf13fb3",
+      "fileHash": "be8cf446",
+      "needsInterop": false
+    },
+    "@angular/common/http": {
+      "src": "../../../../../node_modules/@angular/common/fesm2022/http.mjs",
+      "file": "@angular_common_http.js",
+      "fileHash": "dbda909a",
       "needsInterop": false
     },
     "@angular/core": {
       "src": "../../../../../node_modules/@angular/core/fesm2022/core.mjs",
       "file": "@angular_core.js",
-      "fileHash": "16d09a59",
+      "fileHash": "19b375df",
       "needsInterop": false
     },
     "@angular/forms": {
       "src": "../../../../../node_modules/@angular/forms/fesm2022/forms.mjs",
       "file": "@angular_forms.js",
-      "fileHash": "b776a409",
+      "fileHash": "47c0827e",
       "needsInterop": false
     },
     "@angular/material/button": {
       "src": "../../../../../node_modules/@angular/material/fesm2022/button.mjs",
       "file": "@angular_material_button.js",
-      "fileHash": "b77ab62b",
+      "fileHash": "55aaabe1",
       "needsInterop": false
     },
     "@angular/material/form-field": {
       "src": "../../../../../node_modules/@angular/material/fesm2022/form-field.mjs",
       "file": "@angular_material_form-field.js",
-      "fileHash": "63894f61",
+      "fileHash": "09887fbe",
       "needsInterop": false
     },
     "@angular/material/input": {
       "src": "../../../../../node_modules/@angular/material/fesm2022/input.mjs",
       "file": "@angular_material_input.js",
-      "fileHash": "cd9349ff",
+      "fileHash": "574234b6",
       "needsInterop": false
     },
     "@angular/material/select": {
       "src": "../../../../../node_modules/@angular/material/fesm2022/select.mjs",
       "file": "@angular_material_select.js",
-      "fileHash": "335ad607",
+      "fileHash": "e730ec42",
       "needsInterop": false
     },
     "@angular/platform-browser": {
       "src": "../../../../../node_modules/@angular/platform-browser/fesm2022/platform-browser.mjs",
       "file": "@angular_platform-browser.js",
-      "fileHash": "110eb004",
+      "fileHash": "ef20e243",
       "needsInterop": false
     },
     "@angular/platform-browser/animations/async": {
       "src": "../../../../../node_modules/@angular/platform-browser/fesm2022/animations/async.mjs",
       "file": "@angular_platform-browser_animations_async.js",
-      "fileHash": "ae490343",
+      "fileHash": "2071ddbb",
       "needsInterop": false
     },
     "@angular/router": {
       "src": "../../../../../node_modules/@angular/router/fesm2022/router.mjs",
       "file": "@angular_router.js",
-      "fileHash": "4c1151b3",
+      "fileHash": "bde3e43d",
       "needsInterop": false
     },
     "rxjs": {
       "src": "../../../../../node_modules/rxjs/dist/esm5/index.js",
       "file": "rxjs.js",
-      "fileHash": "12061d0c",
+      "fileHash": "11807b4b",
       "needsInterop": false
     }
   },
@@ -75,11 +81,14 @@
     "browser-UOK26EQR": {
       "file": "browser-UOK26EQR.js"
     },
-    "chunk-4NJKIFMT": {
-      "file": "chunk-4NJKIFMT.js"
-    },
     "chunk-KBCLPT66": {
       "file": "chunk-KBCLPT66.js"
+    },
+    "chunk-ZN4ZQIKF": {
+      "file": "chunk-ZN4ZQIKF.js"
+    },
+    "chunk-BBUMVJIZ": {
+      "file": "chunk-BBUMVJIZ.js"
     },
     "chunk-RGXM7TCY": {
       "file": "chunk-RGXM7TCY.js"
