+++ Aulas-Angular/app-modules/src/app/modules/public/home/home.component.html
@@ -1,57 +1,80 @@
-
 <div class="pai">
-<div class="cont">
-<h1>Conheça o TaskMaster</h1>
-
-</div>
-<div class="txt">
+  <div class="cont">
+    <h1>Conheça o TaskMaster</h1>
+  </div>
+  <div class="txt">
     <div class="txt2">
-        <h2> Aprenda a controlar sua rotina no cotidiano </h2>
-    </div>
-<h4>
-    Seu Novo Parceiro em Produtividade
-Seja bem-vindo ao TaskMaster, a solução definitiva para quem busca organização e eficiência no dia a dia. Nossa plataforma foi desenvolvida com um único objetivo: simplificar a gestão de tarefas, tornando sua vida mais produtiva e menos estressante.
-</h4>
-<!--  -->
-
-</div>
-<div class="txt2">
-    <h2> Por que escolher o TaskMaster? </h2>
-</div>
-<h4>
-    <ul>
-        <li>
-            Interface Intuitiva e Amigável: Nosso design limpo e intuitivo permite que você comece a organizar suas tarefas em segundos, sem necessidade de aprendizado complicado.
-        </li>
-        <li>
-            Gerenciamento Completo de Tarefas: Crie, visualize, edite e priorize suas tarefas de maneira fácil e rápida. Classifique por data, prioridade ou projeto para manter o foco no que realmente importa.
-        </li>
-        <li>
-            Colaboração Simplificada: Compartilhe listas de tarefas com colegas de trabalho, amigos ou familiares. Facilite a colaboração e mantenha todos na mesma página.
-        </li>
-        <li>
-            Notificações e Lembretes: Nunca mais perca um prazo importante. Receba lembretes e notificações para suas tarefas diretamente em seu dispositivo.
-        </li>
-        <li>
-            Acessível em Todos os Dispositivos: Acesse suas listas de tarefas de qualquer lugar, a qualquer momento, seja no seu desktop, tablet ou smartphone.
-        </li>
-        
-            <!-- -->
-    </ul>
-    </h4>
-    <div class="txt2">
-        <h2>Cadastre-se</h2> 
+      <h2>Aprenda a controlar sua rotina no cotidiano</h2>
     </div>
     <h4>
-        Experimente o TaskMaster hoje mesmo e transforme sua maneira de gerenciar tarefas. Cadastre-se gratuitamente em nosso site e descubra como podemos ajudá-lo a alcançar novos níveis de produtividade e organização. Não deixe para depois, comece a simplificar sua vida agora com TaskMaster!
+      Seu Novo Parceiro em Produtividade Seja bem-vindo ao TaskMaster, a solução
+      definitiva para quem busca organização e eficiência no dia a dia. Nossa
+      plataforma foi desenvolvida com um único objetivo: simplificar a gestão de
+      tarefas, tornando sua vida mais produtiva e menos estressante.
     </h4>
+    <!--  -->
+  </div>
+  <div class="txt2">
+    <h2>Por que escolher o TaskMaster?</h2>
+  </div>
+  <h4>
+    <ul>
+      <li>
+        Interface Intuitiva e Amigável: Nosso design limpo e intuitivo permite
+        que você comece a organizar suas tarefas em segundos, sem necessidade de
+        aprendizado complicado.
+      </li>
+      <li>
+        Gerenciamento Completo de Tarefas: Crie, visualize, edite e priorize
+        suas tarefas de maneira fácil e rápida. Classifique por data, prioridade
+        ou projeto para manter o foco no que realmente importa.
+      </li>
+      <li>
+        Colaboração Simplificada: Compartilhe listas de tarefas com colegas de
+        trabalho, amigos ou familiares. Facilite a colaboração e mantenha todos
+        na mesma página.
+      </li>
+      <li>
+        Notificações e Lembretes: Nunca mais perca um prazo importante. Receba
+        lembretes e notificações para suas tarefas diretamente em seu
+        dispositivo.
+      </li>
+      <li>
+        Acessível em Todos os Dispositivos: Acesse suas listas de tarefas de
+        qualquer lugar, a qualquer momento, seja no seu desktop, tablet ou
+        smartphone.
+      </li>
 
-<div class="img">
-    <img class="seta" src="../../../../assets/setaroxa.png" alt="">
-    <div class="btn"><button mat-raised-button *ngIf="!this.router.url.includes('/signup')" routerLink="/signup">Inscreva-se</button></div>
-</div>
+      <!-- -->
+    </ul>
+  </h4>
+  <div class="txt2">
+    <h2>Cadastre-se</h2>
+  </div>
+  <h4>
+    Experimente o TaskMaster hoje mesmo e transforme sua maneira de gerenciar
+    tarefas. Cadastre-se gratuitamente em nosso site e descubra como podemos
+    ajudá-lo a alcançar novos níveis de produtividade e organização. Não deixe
+    para depois, comece a simplificar sua vida agora com TaskMaster!
+  </h4>
 
-<h4>    
-    Junte-se a milhares de usuários que estão aproveitando ao máximo cada dia com nosso sistema de gerenciamento de tarefas. Não espere mais para transformar sua produtividade e eficiência. Estamos aqui para ajudar você a conquistar mais, uma tarefa de cada vez. 
-</h4>
+  <div class="img">
+    <img class="seta" src="../../../../assets/setaroxa.png" alt="" />
+    <div class="btn">
+      <button
+        mat-raised-button
+        *ngIf="!this.router.url.includes('/signup')"
+        routerLink="/signup"
+      >
+        Inscreva-se
+      </button>
+    </div>
+  </div>
+
+  <h4>
+    Junte-se a milhares de usuários que estão aproveitando ao máximo cada dia
+    com nosso sistema de gerenciamento de tarefas. Não espere mais para
+    transformar sua produtividade e eficiência. Estamos aqui para ajudar você a
+    conquistar mais, uma tarefa de cada vez.
+  </h4>
 </div>
