+++ Aulas-Angular/app-modules/src/app/modules/shared/services/auth.service.ts
@@ -1,9 +1,26 @@
+import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
+import { apiUrl } from '../../../../environments/environment';
 
+type TUser = {
+  email: string,
+  password: string
+}
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class AuthService {
+  header = {
+    'Content-Type': 'application/json',
+    Authorization: `Bearer ${localStorage.getItem('access_token')}`,
+  };
+  constructor(private http: HttpClient) {}
 
-  constructor() { }
+  public authenticate(credentials: TUser) {
+    return this.http.post(
+      `${apiUrl}/auth/authenticate`,
+      credentials,
+      { headers: this.header }
+    );
+  }
 }
